Finished the implementation of the Supplier Invoice page
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C26197

+++ TOneV2/Code/TOneV2/TOne.Billing.Entities/SupplierInvoiceDetail.cs
@@ -8,7 +8,7 @@ namespace TOne.Billing.Entities
 {
     public class SupplierInvoiceDetail
     {
-        public int DetailID { get; set; }
+        public long DetailID { get; set; }
         public DateTime FromDate { get; set; }
         public DateTime TillDate { get; set; }
         public string Destination { get; set; }
